Added function and tests for calculating the specific heat capacity with the formula derived earlier.
